# docs/api-docs-parquimetro-api - tech-challenge-fase-2.json
{"openapi":"3.0.1","info":{"title":"Parquímetro API","description":"API desenvolvida com o objetivo de reescrever um sistema de parquímetro visando otimização e alto desempenho durante o Tech Challenge da Fase 2 na Pos Tech FIAP em Arquitetura e Desenvolvimento Java.","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Parquímetro [ParkingController]","description":"Controlador responsavel por gerenciar informações gerais do sistema de parquímetro."},{"name":"Veículos [VehicleController]","description":"Controlador que administra dados dos veículos no sistema de parquímetro."},{"name":"Agendamentos [SchedulingController]","description":"Controlador encarregado de fornecer requisições associadas ao agendamento no sistema de parquímetro."}],"paths":{"/vehicles":{"get":{"tags":["Veículos [VehicleController]"],"summary":"Enpoint para listar veículos cadastrados.","operationId":"findAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Vehicle"}}}}}}},"put":{"tags":["Veículos [VehicleController]"],"summary":"Enpoint para atualizar informações de um veículos cadastrados anteriormente.","operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicle"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}},"post":{"tags":["Veículos [VehicleController]"],"summary":"Recurso para registrar veículos no sistema.","operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicle"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}}},"/schedulings":{"get":{"tags":["Agendamentos [SchedulingController]"],"summary":"Listagem de todos os agendamentos no parquimetro.","operationId":"findAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Scheduling"}}}}}}},"put":{"tags":["Agendamentos [SchedulingController]"],"summary":"Recurso para atualização de agendamentos.","operationId":"update_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Scheduling"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Scheduling"}}}}}},"post":{"tags":["Agendamentos [SchedulingController]"],"summary":"Endpoint para registrar o agendamento no parquimetro.","operationId":"save_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Scheduling"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Scheduling"}}}}}}},"/vehicles/{id}":{"get":{"tags":["Veículos [VehicleController]"],"summary":"Funcionalidade para regatar um veículos cadastrados pela placa.","operationId":"findBy","parameters":[],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}}},"/schedulings/{plate}/status":{"get":{"tags":["Agendamentos [SchedulingController]"],"summary":"Verificação de status baseado no último agendamento do parquimetro através da placa.","operationId":"checkSchedulingStatusByPlate","parameters":[{"name":"plate","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SchedulingStatusDTO"}}}}}}},"/schedulings/{id}":{"get":{"tags":["Agendamentos [SchedulingController]"],"summary":"Listagem do agendamento no parquimetro por id.","operationId":"findBy_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Scheduling"}}}}}}},"/parkings/calculate":{"get":{"tags":["Parquímetro [ParkingController]"],"summary":"Retona o calculo do valor total a pagar pelas horas reservadas.","operationId":"calculateParkingByHour","parameters":[{"name":"hours","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"number"}}}}}}}},"components":{"schemas":{"Conductor":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"cpf":{"type":"string"},"email":{"type":"string"},"cellphone":{"type":"string"}}},"Vehicle":{"type":"object","properties":{"plate":{"type":"string"},"conductor":{"$ref":"#/components/schemas/Conductor"}}},"Payment":{"type":"object","properties":{"value":{"type":"number"},"type":{"type":"string","enum":["CASH","CREDIT","DEBIT","CRIPTO","PIX"]},"realizationDate":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["OK","NOT_OK"]}}},"Scheduling":{"type":"object","properties":{"id":{"type":"string"},"contractedHours":{"type":"integer","format":"int32"},"startDate":{"type":"string","format":"date-time"},"vehiclePlate":{"type":"string"},"payment":{"$ref":"#/components/schemas/Payment"},"version":{"type":"integer","format":"int64"}}},"SchedulingStatusDTO":{"type":"object","properties":{"status":{"type":"string","enum":["OK","NOT_SCHEDULED","EXPIRED"]},"endsAt":{"type":"string","format":"date-time"}}}}}}