# src/main/java/br/com/fiap/equipe3/secondtechchallenge/models/Vehicle.java
package br.com.fiap.equipe3.secondtechchallenge.models;

import lombok.Data;
import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.DBRef;
import org.springframework.data.mongodb.core.mapping.Document;

@Data
@Document(collection = "vehicles")
public class Vehicle {
    @Id
    private String plate;

    @DBRef
    private Conductor conductor;
}